# Conflicts:
#	resources/images/terrain/biome_palettes/palette_clear+aw2.png
#	resources/images/terrain/biome_palettes/palette_clear+awds.png
#	resources/images/terrain/biome_palettes/palette_clear+dor.png
#	resources/images/terrain/biome_palettes/palette_clear.png
#	resources/images/terrain/biome_palettes/palette_desert+awds.png
#	resources/images/terrain/biome_palettes/palette_desert+dor.png
#	resources/images/terrain/biome_palettes/palette_desert.png
#	resources/images/terrain/biome_palettes/palette_guide.png
#	resources/images/terrain/biome_palettes/palette_rain+aw2.png
#	resources/images/terrain/biome_palettes/palette_snow+aw2.png
#	resources/images/terrain/biome_palettes/palette_snow+awds.png
#	resources/images/terrain/biome_palettes/palette_snow+dor.png
#	resources/images/terrain/biome_palettes/palette_snow.png
#	resources/images/terrain/biome_palettes/palette_waste+awds.png
#	resources/images/terrain/biome_palettes/palette_waste+dor.png
#	resources/images/terrain/biome_palettes/palette_waste.png

+++ 3rd_party/oxygine-framework/oxygine/actor/slidingsprite.h
@@ -12,10 +12,10 @@ namespace oxygine
         explicit SlidingSprite();
         virtual ~SlidingSprite() = default;
         void setResAnim(ResAnim* pAnim);
-        float getSpeedX() const;
-        void setSpeedX(float speed);
-        float getSpeedY() const;
-        void setSpeedY(float speed);
+        qint32 getSpeedX() const;
+        void setSpeedX(qint32 speed);
+        qint32 getSpeedY() const;
+        void setSpeedY(qint32 speed);
 
         bool getFlippedX() const;
         void setFlippedX(bool flipped);
@@ -25,8 +25,8 @@ namespace oxygine
 
     private:
         oxygine::spSprite m_sprites[4];
-        float m_speedX{1};
-        float m_speedY{0};
+        qint32 m_speedX{1};
+        qint32 m_speedY{0};
         bool m_flippedX{false};
         bool m_locked{false};
     };
